Fixes bugs related to pagination on search and row count
When searching after you've already paginated, if the results of the search provide fewer results than would exist in the current pagination, it would show zero records. The same is true for changing the number of rows to display after pagination. These two line changes force a page refresh whenever a search or row count is updated, ensuring we go back to page 1 of pagination